refactor tests and use consistent namings
