+++ examples/custom-pipelines-hello-world/src/main/java/co/decodable/examples/cpdemo/KeyedTableAPIJob.java
@@ -7,7 +7,7 @@
  */
 package co.decodable.examples.cpdemo;
 
-import co.decodable.examples.cpdemo.model.KeyedPurchaseOrder;
+import co.decodable.examples.cpdemo.model.append.KeyedPurchaseOrder;
 import co.decodable.examples.cpdemo.model.OrderKey;
 import co.decodable.examples.cpdemo.model.PurchaseOrder;
 import co.decodable.sdk.pipeline.DecodableStreamSink;
@@ -53,7 +53,7 @@ public class KeyedTableAPIJob {
         env.fromSource(
             source,
             WatermarkStrategy.noWatermarks(),
-            "[stream-purchase-orders] Purchase Orders Source");
+            PURCHASE_ORDERS_STREAM);
 
     Table inputTable = tableEnv.fromDataStream(
         stream,
@@ -104,9 +104,9 @@ public class KeyedTableAPIJob {
             )
         );
 
-    resultStream.sinkTo(sink).name("[stream-purchase-orders-processed] Purchase Orders Sink");
+    resultStream.sinkTo(sink).name(PURCHASE_ORDERS_PROCESSED_STREAM);
 
-    env.execute("Purchase Order Processor");
+    env.execute("purchase order processor with keyed append streams");
   }
 
   // UDF
